v0.3 [int]GraphNode ID added,which is:     [standalone]Memory address,     [in manager]vector address. init in-built binary tree support [Dot file generate] support undirected graph and binary tree

+++ CustomDataStructureTest/CDST.h
@@ -8,9 +8,9 @@ public:
 	Linker();
 	Linker(int);
 	PVOID Port1;
-	int Id1;
+	size_t Id1;
 	PVOID Port2;
-	int Id2;
+	size_t Id2;
 	int Weight;
 	PVOID GetThat(PVOID);
 }*pLinker;
